+++ CHANGES.md
@@ -3,6 +3,7 @@ Changes in RiotX 0.23.0 (2020-XX-XX)
 
 Features ✨:
  - Call with WebRTC support (##611)
+ - Add capability to change the display name (#1529)
 
 Improvements 🙌:
  - "Add Matrix app" menu is now always visible (#1495)